MenuGroupsPopup component extraction
