+++ app/src/main/java/slowscript/warpinator/feature/home/components/HomeMenu.kt
@@ -3,8 +3,6 @@ package slowscript.warpinator.feature.home.components
 import android.app.Activity
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.material.icons.Icons
@@ -17,16 +15,9 @@ import androidx.compose.material.icons.rounded.MoreVert
 import androidx.compose.material.icons.rounded.Refresh
 import androidx.compose.material.icons.rounded.Settings
 import androidx.compose.material.icons.rounded.WifiTethering
-import androidx.compose.material3.DropdownMenuGroup
-import androidx.compose.material3.DropdownMenuItem
-import androidx.compose.material3.DropdownMenuPopup
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.MenuDefaults
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
@@ -35,14 +26,13 @@ import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import slowscript.warpinator.app.LocalNavController
+import slowscript.warpinator.core.design.components.MenuAction
+import slowscript.warpinator.core.design.components.MenuGroup
+import slowscript.warpinator.core.design.components.MenuGroupsPopup
 import slowscript.warpinator.core.design.components.TooltipIconButton
-
-data class MenuAction(val title: String, val icon: ImageVector, val onClick: () -> Unit)
-data class MenuGroup(val actions: List<MenuAction>)
 
 
 @OptIn(ExperimentalMaterial3ExpressiveApi::class, ExperimentalMaterial3Api::class)
@@ -117,38 +107,12 @@ fun HomeMenu(
             icon = Icons.Rounded.MoreVert,
             onClick = { menuOpen = true },
         )
-        DropdownMenuPopup(
-            expanded = menuOpen,
-            onDismissRequest = { menuOpen = false },
-
-            ) {
-            menuGroups.forEachIndexed { index, group ->
-                DropdownMenuGroup(
-                    shapes = MenuDefaults.groupShape(index, menuGroups.size),
-                    interactionSource = groupInteractionSource,
-                    containerColor = MaterialTheme.colorScheme.secondaryContainer,
-                ) {
-                    group.actions.forEach { action ->
-                        DropdownMenuItem(
-                            text = { Text(action.title) },
-                            trailingIcon = {
-                                Icon(
-                                    action.icon, contentDescription = null,
-                                )
-                            },
-                            onClick = {
-                                action.onClick()
-                                menuOpen = false
-                            },
-                        )
-                    }
-
-                }
-                if (index != menuGroups.size - 1) {
-                    Spacer(Modifier.height(MenuDefaults.GroupSpacing))
-                }
-            }
-        }
+        MenuGroupsPopup(
+            menuGroups = menuGroups,
+            groupInteractionSource = groupInteractionSource,
+            menuOpen = menuOpen,
+            onDismiss = { menuOpen = false },
+        )
     }
 }
